moved the AddEventForm to correct folder
